fix: improve data validation
